refactor: reorganize project directory structure - Move public headers to include/
