+++ containers/internal/stack_internal_data.h
