+++ core/internal/core_string_internal_data.h
