+++ include/core/core_memory.h
