+++ include/core/core_string.h
