+++ include/core/message.h
